requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# PyCrack/main.py: 3,4
pyelftools == 0.27
